requirements.txt
torch>=2.0.0
diffusers>=0.25.0
transformers>=4.38.0
accelerate>=0.27.0
datasets>=2.16.0
Pillow>=9.0.0
numpy>=1.24.0
tqdm>=4.65.0
pyyaml>=6.0
wandb>=0.16.0
